+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Add exposure time and number of exposure per tile to the Overwatcher configuration.
 * Do not try to emit Slack notifications if the internet is down.
+* Allow to specify the default exposure time for the AG cameras in the configuration file.
 
 ### 🔧 Fixed